fix(torghut): clear doc71 autoresearch ci gates
