+++ docs/jangar/architecture-inventory.md
@@ -29,9 +29,9 @@ This file is generated by `bun run --cwd services/jangar docs:inventory:write`.
 | 13   | `services/jangar/src/server/agents-controller/agent-run-reconciler.ts` | 1191 |
 | 14   | `services/jangar/src/server/agents-controller/workflow-reconciler.ts`  | 1191 |
 | 15   | `services/jangar/src/server/agentctl-grpc.ts`                          | 1171 |
-| 16   | `services/jangar/src/server/torghut-trading.ts`                        | 1081 |
-| 17   | `services/jangar/src/routes/github/pulls/$owner/$repo/$number.tsx`     | 1028 |
-| 18   | `services/jangar/src/server/db.ts`                                     | 1017 |
+| 16   | `services/jangar/src/routes/github/pulls/$owner/$repo/$number.tsx`     | 1028 |
+| 17   | `services/jangar/src/server/db.ts`                                     | 1017 |
+| 18   | `services/jangar/src/server/torghut-trading.ts`                        |  988 |
 | 19   | `services/jangar/src/routes/torghut/trading.tsx`                       |  980 |
 | 20   | `services/jangar/src/server/agents-controller/vcs-context.ts`          |  975 |
